+++ Cmdr/BuiltInCommands/Debug/fetch.lua
@@ -1,0 +1,13 @@
+return {
+	Name = "fetch";
+	Aliases = {};
+	Description = "Fetch a value from the Internet";
+	Group = "DefaultDebug";
+	Args = {
+		{
+			Type = "url";
+			Name = "URL";
+			Description = "The URL to fetch.";
+		}
+	};
+}