+++ Cmdr/BuiltInCommands/Debug/fetchServer.lua
@@ -1,0 +1,5 @@
+local HttpService = game:GetService("HttpService")
+
+return function (_, url)
+	return HttpService:GetAsync(url)
+end